requirements.txt
altair==5.5.0
attrs==25.1.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
fasttext==0.9.3
gcloud==0.18.3
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.25.1
numpy==1.26.4
oauth2client==4.1.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
plotly==6.0.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PyPDF2==3.0.1
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.37
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
watchdog==6.0.0
